+++ ea_airflow_util/dags/run_dbt_airflow_dag.py
@@ -61,7 +61,7 @@ class RunDbtDag():
         slack_conn_id: Optional[str] = None,
         dbt_incrementer_var: str = None,
 
-        trigger_dag_on_run_success: Optional[str] = None,
+        trigger_dags_on_run_success: Optional[list] = None,
 
         **kwargs
     ):
@@ -120,15 +120,19 @@ class RunDbtDag():
             self.dbt_var_reset_operator = None
 
         # Build optional operator to trigger downstream DAG when `dbt run` succeeds.
-        if trigger_dag_on_run_success:
-            self.trigger_dag_on_run_success_operator = TriggerDagRunOperator(
-                task_id="trigger_downstream_dag",
-                trigger_dag_id=trigger_dag_on_run_success,
-                wait_for_completion=False,  # Keep running DBT DAG while downstream DAG runs.
-                trigger_rule='all_success',
-            )
+        if trigger_dags_on_run_success:
+            self.external_dags = []
+            
+            for external_dag_id in trigger_dags_on_run_success:
+                self.external_dags.append(
+                    TriggerDagRunOperator(
+                        task_id=f"trigger_{external_dag_id}",
+                        trigger_dag_id=external_dag_id,
+                        wait_for_completion=False,  # Keep running DBT DAG while downstream DAG runs.
+                        trigger_rule='all_success',
+                ))
         else:
-            self.trigger_dag_on_run_success_operator = None
+            self.external_dags = None
 
 
     # create DAG 
@@ -244,8 +248,8 @@ class RunDbtDag():
                 dbt_build_artifact_tables >> dbt_seed
 
             # Trigger downstream DAG when `dbt run` succeeds
-            if self.trigger_dag_on_run_success_operator:
-                dbt_run >> self.trigger_dag_on_run_success_operator
+            if self.external_dags:
+                dbt_run >> self.external_dags
 
         # Apply the DBT variable operators if defined.
         if self.dbt_incrementer_var: